Add files via upload
In this project, we have used k-means clustering to cluster the existing users based on their locations. The new user provides its location as well as its preferred cuisine according to which the recommender system assigns the user to one of the clusters and suggest a restaurant based on its cuisine preference and mean restaurant rating.